Update test_AR.py
Loop function works correctly. Just set the amount of time and wells with sample at the beginning. In order to save each image with the number of the well, a list with all the wells filled must be generated by hand at the beginning.  Added to the loop the path and the name of the image to be saved, with the number and the amount of time run since the start of the experiment